Changed paradigm for drawBars to difference between 2 arrays
Instead of highlighting a specific bar according to a position included
in an array. I pass 2 array and highlights the bar if there is a
difference between the 2 arrays at that specific position.

The function also updates the second array with the first one, so at the
end of the function the 2 array are the same.